+++ lib/questions_screen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 
 import 'package:quiz_app/answer_button.dart';
-import 'package:quiz_app/data/questions.dart';
+import 'package:quiz_app/data/flutter_questions.dart';
 
 class QuestionsScreen extends StatefulWidget {
   const QuestionsScreen({super.key, required this.onSelectAnswer});
